Get timeslots working for counter
